preparation for second container

+++ deploy_database.sh
@@ -1,5 +1,10 @@
+if docker inspect --format '{{json .State.Running}}' mongo_database
+then
+  echo "container is already running"
+  exit
+fi
 docker pull mongo
 docker rm -f mongo_database || echo no service to remove
 
 docker run -d --name mongo_database -p 27017:27017 -v mongodb_data:/data/db mongo
-docker logs -f mongo_database
+echo 'database launched successfully'
